+++ Breakin/src/Engine/GameTime.cpp
@@ -1,4 +1,4 @@
-#include "gametime.h"
+#include "GameTime.h"
 
 #include <algorithm>
 #include <chrono>
@@ -11,17 +11,17 @@ void bin::GameTime::Increment()
 
     // Calculate delta time
     const auto currentTime = std::chrono::high_resolution_clock::now();
-    g_DeltaTime = std::min(MAX_DELTA_TIME, std::chrono::duration<double>(currentTime - s_LastTime).count());
+    g_DeltaTime = std::min(MAX_DELTA_TIME, std::chrono::duration<float>(currentTime - s_LastTime).count());
     s_LastTime = currentTime;
 
     // Update elapsed time
     g_ElapsedTime += GetDeltaTime();
 }
 
-double bin::GameTime::GetDeltaTime() { return g_DeltaTime * g_TimeScale; }
+float bin::GameTime::GetDeltaTime() { return g_DeltaTime * g_TimeScale; }
 
-double bin::GameTime::GetUnscaledDeltaTime() { return g_DeltaTime; }
+float bin::GameTime::GetUnscaledDeltaTime() { return g_DeltaTime; }
 
-double bin::GameTime::GetFixedDeltaTime() { return g_FixedDeltaTime; }
+float bin::GameTime::GetFixedDeltaTime() { return g_FixedDeltaTime; }
 
-double bin::GameTime::GetElapsedTime() { return g_ElapsedTime; }
+float bin::GameTime::GetElapsedTime() { return g_ElapsedTime; }
